requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
drf-yasg==1.21.8
psycopg2-binary==2.9.10
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
gunicorn==23.0.0
dj-database-url==2.3.0
whitenoise==6.8.2
